Create last-resorts and transpositions at build time
