+++ last-resorts.json
@@ -1,0 +1,1 @@
+[["MIT +NO-FALSE-ATTRIBS","MITNFA"],["LZMA-SDK-9.11-to","LZMA-SDK-9.11-to-9.20"],["Community-Spec","Community-Spec-1.0"],["CDLA-Sharing","CDLA-Sharing-1.0"],["Hippocratic","Hippocratic-2.1"],["LiLiQ-Rplus","LiLiQ-Rplus-1.1"],["OGDL-Taiwan","OGDL-Taiwan-1.0"],["CERN-OHL-P","CERN-OHL-P-2.0"],["CERN-OHL-S","CERN-OHL-S-2.0"],["CERN-OHL-W","CERN-OHL-W-2.0"],["DL-DE-ZERO","DL-DE-ZERO-2.0"],["libselinux","libselinux-1.0"],["OGL-Canada","OGL-Canada-2.0"],["Frameworx","Frameworx-1.0"],["Inner-Net","Inner-Net-2.0"],["Interbase","Interbase-1.0"],["2 CLAUSE","BSD-2-Clause"],["2-CLAUSE","BSD-2-Clause"],["3 CLAUSE","BSD-3-Clause"],["3-CLAUSE","BSD-3-Clause"],["ARTISTIC","Artistic-2.0"],["DL-DE-BY","DL-DE-BY-2.0"],["LZMA-SDK","LZMA-SDK-9.22"],["Sendmail","Sendmail-8.23"],["TAPR-OHL","TAPR-OHL-1.0"],["CUA-OPL","CUA-OPL-1.0"],["ECLIPSE","EPL-1.0"],["Elastic","Elastic-2.0"],["LiLiQ-P","LiLiQ-P-1.1"],["LiLiQ-R","LiLiQ-R-1.1"],["NCGL-UK","NCGL-UK-2.0"],["OpenPBS","OpenPBS-2.3"],["OSET-PL","OSET-PL-2.1"],["Unicode","Unicode-3.0"],["XFree86","XFree86-1.1"],["AFFERO","AGPL-3.0-or-later"],["Affero","AGPL-3.0-or-later"],["APACHE","Apache-2.0"],["CATOSL","CATOSL-1.1"],["Condor","Condor-1.1"],["etalab","etalab-2.0"],["JasPer","JasPer-2.0"],["libpng","libpng-2.0"],["ODC-By","ODC-By-1.0"],["OPL-UK","OPL-UK-3.0"],["Python","Python-2.0"],["RHeCos","RHeCos-1.1"],["SAX-PD","SAX-PD-2.0"],["TORQUE","TORQUE-1.1"],["Watcom","Watcom-1.0"],["Xdebug","Xdebug-1.03"],["BOOST","BSL-1.0"],["C-UDA","C-UDA-1.0"],["D-FSL","D-FSL-1.0"],["ErlPL","ErlPL-1.1"],["GPL-1","GPL-1.0-only"],["GPL-2","GPL-2.0-only"],["GPLV1","GPL-1.0-only"],["GPLV2","GPL-2.0-only"],["NICTA","NICTA-1.0"],["NPOSL","NPOSL-3.0"],["O-UDA","O-UDA-1.0"],["OPUBL","OPUBL-1.0"],["SimPL","SimPL-2.0"],["SISSL","SISSL-1.2"],["TGPPL","TGPPL-1.0"],["AGPL","AGPL-3.0-or-later"],["BEER","Beerware"],["BUSL","BUSL-1.1"],["CDDL","CDDL-1.1"],["COIL","COIL-1.0"],["CPAL","CPAL-1.0"],["CPOL","CPOL-1.02"],["FUCK","WTFPL"],["LGPL","LGPL-3.0-or-later"],["NASA","NASA-1.3"],["NBPL","NBPL-1.0"],["OCLC","OCLC-2.0"],["ODbL","ODbL-1.0"],["OLFL","OLFL-1.3"],["PDDL","PDDL-1.0"],["RPSL","RPSL-1.0"],["SSPL","SSPL-1.0"],["UNLI","Unlicense"],["Zend","Zend-2.0"],["ZLIB","Zlib"],["BSD","BSD-2-Clause"],["BSL","BSL-1.0"],["CAL","CAL-1.0"],["CC0","CC0-1.0"],["CDL","CDL-1.0"],["CPL","CPL-1.0"],["GNU","GPL-3.0-or-later"],["GPL","GPL-3.0-or-later"],["IPL","IPL-1.0"],["MIT","MIT"],["MPL","MPL-2.0"],["OGC","OGC-1.0"],["OPL","OPL-1.0"],["PSF","PSF-2.0"],["QPL","QPL-1.0"],["SPL","SPL-1.0"],["TPL","TPL-1.0"],["UCL","UCL-1.0"],["UPL","UPL-1.0"],["VSL","VSL-1.0"],["WTF","WTFPL"],["X11","X11"]]
