+++ transpositions.json
@@ -1,0 +1,1 @@
+[["GNU LESSER GENERAL PUBLIC LICENSE","LGPL"],["GNU Lesser General Public License","LGPL"],["GNU LESSER GENERAL PUBLIC LICENSE","LGPL-2.1"],["GNU Lesser General Public License","LGPL-2.1"],["LESSER GENERAL PUBLIC LICENSE","LGPL"],["Lesser General Public License","LGPL"],["LESSER GENERAL PUBLIC LICENSE","LGPL-2.1"],["Lesser General Public License","LGPL-2.1"],["Universal Permissive License","UPL"],["GNU General Public License","GPL"],["GNU GENERAL PUBLIC LICENSE","GPL"],["Mozilla Public License","MPL"],["Gnu public license","GPL"],["GNU Public License","GPL"],[" International",""],[" or later","+"],["-License",""],["GNU LGPL","LGPL"],["GNU GLP","GPL"],["GNU GPL","GPL"],["GNU/GPL","GPL"],["Claude","Clause"],["WTFGPL","WTFPL"],["APGL","AGPL"],["APL","Apache"],["GLP","GPL"],["GLP","GPL"],["GNU","GPL"],["Gpl","GPL"],["GUN","GPL"],["ISD","ISC"],["IST","ISC"],["MTI","MIT"],["WTH","WTF"],["+",""]]
